accelerometer-plotting/requirements.txt
autopep8==1.5.2
cycler==0.10.0
kiwisolver==1.2.0
matplotlib==3.2.1
numpy==1.18.3
pycodestyle==2.5.0
pyparsing==2.4.7
python-dateutil==2.8.1
six==1.14.0
